#10 submenus de insertar de procesadores y placa base

+++ DB_Procesadores_Java/src/main/java/es/iespuertodelacruz/procesadores/api/NombreProcesador.java
@@ -12,7 +12,7 @@ public class NombreProcesador {
      * @param familia a la que pertenece
      * @param generacion a la que pertenece
      */
-    public NombreProcesador(String modeloProcesador, String familia, byte generacion) {
+    public NombreProcesador(String modeloProcesador, String familia, int generacion) {
         this.modeloProcesador = modeloProcesador;
         this.familia = familia;
         this.generacion = generacion;
